Simple replay is successfully parsed!

+++ src/main/java/com/w3gdata/actionblock/ChangeAllyOptions.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class ChangeAllyOptions implements ActionBlock {
-    public static final byte ID = 0x50;
+    public static final int ID = 0x50;
 
     public int playerSlotNumber;
