+++ src/main/java/com/w3gdata/actionblock/ContinueGameB.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class ContinueGameB implements Unknown {
-    public static final byte ID = 0x69;
+    public static final int ID = 0x69;
 
     public static final int ADDITIONAL_BYTE = 0x68;//for WarCraft III patch version <= 1.06
