+++ src/main/java/com/w3gdata/actionblock/EnterChooseBuildingSubmenu.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class EnterChooseBuildingSubmenu implements ActionBlock {
-    public static final byte ID = 0x67;
+    public static final int ID = 0x67;
 
     public static final int ADDITIONAL_BYTE = 0x66;//for WarCraft III patch version <= 1.06
 }
