+++ src/main/java/com/w3gdata/actionblock/EscPressed.java
@@ -1,5 +1,5 @@
 package com.w3gdata.actionblock;
 
 public class EscPressed implements ActionBlock {
-    public static final byte ID = 0x61;
+    public static final int ID = 0x61;
 }
