+++ src/main/java/com/w3gdata/actionblock/GameSpeedDecreasing.java
@@ -1,5 +1,5 @@
 package com.w3gdata.actionblock;
 
 public class GameSpeedDecreasing implements ActionBlock {
-    public static final byte ID = 0x05;
+    public static final int ID = 0x05;
 }
