+++ src/main/java/com/w3gdata/actionblock/MapTriggerChatCommand.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class MapTriggerChatCommand implements ActionBlock {
-    public static final byte ID = 0x60;
+    public static final int ID = 0x60;
 
     public int unknownA;
