+++ src/main/java/com/w3gdata/actionblock/SaveGameFinished.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class SaveGameFinished implements ActionBlock {
-    public static final byte ID = 0x07;
+    public static final int ID = 0x07;
 
     public int unknown;//always 0x00000001 so far
