+++ src/main/java/com/w3gdata/actionblock/ScenarioTrigger.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class ScenarioTrigger implements ActionBlock {
-    public static final byte ID = 0x62;
+    public static final int ID = 0x62;
 
     public int unknownA;
