+++ src/main/java/com/w3gdata/actionblock/SelectGroundItem.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class SelectGroundItem implements ActionBlock {
-    public static final byte ID = 0x1C;
+    public static final int ID = 0x1C;
 
     public static final int ADDITIONAL_BYTE = 0x1B;//for WarCraft III patch version <= 1.14b
