+++ src/main/java/com/w3gdata/actionblock/SelectSubgroup.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class SelectSubgroup implements ActionBlock {
-    public static final byte ID = 0x19;
+    public static final int ID = 0x19;
 
     public int subgroupNumber;
