+++ src/main/java/com/w3gdata/actionblock/UnitBuildingAbilityTargetedId.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class UnitBuildingAbilityTargetedId extends UnitBuildingAbilityTargeted {
-    public static final byte ID = 0x012;
+    public static final int ID = 0x012;
 
     public int objId1;
