+++ src/main/java/com/w3gdata/actionblock/Unknown0x21.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class Unknown0x21 implements Unknown {
-    public static final byte ID = 0x21;
+    public static final int ID = 0x21;
 
     @Override
     public int getSize() {
