+++ src/main/java/com/w3gdata/actionblock/Unknown0x75.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class Unknown0x75 implements Unknown {
-    public static int ID = 0x75;
+    public static byte ID = 0x75;
 
     public int unknown;
